backend/requirements.txt
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
asyncpg==0.28.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
coverage==7.13.4
cryptography==46.0.5
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1763918002538/work
fastapi==0.104.1
greenlet @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_greenlet_1771658499/work
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
idna==3.11
iniconfig @ file:///home/conda/feedstock_root/build_artifacts/iniconfig_1760831328738/work
itsdangerous==2.1.2
logging==0.4.9.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079066/work
passlib==1.7.4
pluggy @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pluggy_1764896838/work
prometheus-client==0.17.1
psycopg2-binary==2.9.9
pyasn1==0.6.2
pycparser==3.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
Pygments==2.19.2
PyJWT @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_e55dbbf3h7/croot/pyjwt_1715094746385/work
pytest @ file:///home/conda/feedstock_root/build_artifacts/pytest_1698233724984/work
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.3
rich==13.6.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy @ file:///Users/runner/miniforge3/conda-bld/sqlalchemy_1698950642499/work
starlette==0.27.0
tenacity==8.2.3
tomli @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_tomli_1768146676/work
typing-inspection==0.4.2
typing_extensions==4.8.0
uuid==1.30
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
